// models/request/studentInfoRequest.ts
export interface StudentInfoRequest {
    studentInfo_id: number;
    mobilityType_id: number;
    name: string;
    surname: string;
    gender_id: number;
    nationality_id: number;
    birthdate: string;
    educationTypeAndLevel_id: number;
    email: string;
    subjectArea_id: number;
    subjectAreaDescription: string;
    global_id: string;
    omobility_id: string;
}